complete rework of make script
make now is able to handle both C and Ada files (using a C entrypoint), as well as being able to automatically scan for all files in the /src/ directory recursively and compile them without having to explicitly declare them for compilation. This does mean that test.c has to leave the /src/ directory.